CXP-1289: move attribute model and clean productMappingSchema signature
