+++ components/catalogs/front/src/components/ProductMapping/ProductMapping.integration.tsx
@@ -48,27 +48,14 @@ test('it displays an existing product mapping', async () => {
     };
 
     const productMappingSchema = {
-        $id: 'https://example.com/product',
-        $schema: 'https://api.akeneo.com/mapping/product/0.0.1/schema',
-        $comment: 'My first schema !',
-        title: 'Product Mapping',
-        description: 'JSON Schema describing the structure of products expected by our application',
-        type: 'object',
         properties: {
-            uuid: {
-                type: 'string',
-            },
-            name: {
-                type: 'string',
-            },
+            uuid: {},
+            name: {},
             body_html: {
                 title: 'Description',
-                description: 'Product description in raw HTML',
-                type: 'string',
             },
             erp_name: {
                 title: 'ERP',
-                type: 'string',
             },
         },
     };
