+++ components/catalogs/front/src/components/ProductSelection/hooks/useFindCriterionByField.ts
@@ -1,7 +1,7 @@
 import {useCallback} from 'react';
 import {AnyCriterion} from '../models/Criterion';
 import {useQueryClient} from 'react-query';
-import {Attribute} from '../models/Attribute';
+import {Attribute} from '../../../models/Attribute';
 import {useFindAttributeCriterionByType} from './useFindAttributeCriterionByType';
 import StatusCriterion from '../criteria/StatusCriterion';
 import FamilyCriterion from '../criteria/FamilyCriterion';
