+++ components/catalogs/front/src/components/ProductSelection/hooks/useInfiniteAttributeCriterionFactories.unit.ts
@@ -3,7 +3,7 @@ jest.unmock('./useInfiniteAttributeCriterionFactories');
 import {act, renderHook} from '@testing-library/react-hooks';
 import fetchMock from 'jest-fetch-mock';
 import {mocked} from 'ts-jest/utils';
-import {Attribute} from '../models/Attribute';
+import {Attribute} from '../../../models/Attribute';
 import {ReactQueryWrapper} from '../../../../tests/ReactQueryWrapper';
 import {useInfiniteAttributeCriterionFactories} from './useInfiniteAttributeCriterionFactories';
 import {useFindAttributeCriterionByType} from './useFindAttributeCriterionByType';
