+++ components/catalogs/front/src/models/Attribute.d.ts
